Crate rapi

rapi - A safe and user friendly R extension interface.

This library aims to provide an interface that will be familiar to first-time users of Rust or indeed any compiled language.

Anyone who knows the R library should be able to write R extensions.

This library is just being born, but goals are:

Implement common R functions such as c() and print()

Example:

let v = c!(1, 2, 3);
let l = list!(a=1, b=2);
print!(v, l);

Provide a wrapper for r objects.

Example:

let s = Robj::from("hello");
let i = Robj::from(1);
let r = Robj::from(1.0);

Provide iterator support for creation and consumption of r vectors.

Example:

let res = (1..=100).iter().collect::<Robj>();
for x in res {
    print!(x);
}

Provide a procedural macro to adapt Rust functions to R

Example:

#[derive(RCallable)]
fn fred(a: i32) -> i32 {
    a + 1
}

In R:

 
result <- .Call("fred", 1)
 

Macros

append_lang
Convert a list of tokens to an array of tuples.

args

c
Concatenation operator.

data_frame
Create a dataframe.

lang

list
Create a list.

push_args
Convert a list of tokens to an array of tuples.

r
R object encapsulation operator.

read_table
Read a CSV file.

Structs

Character
Wrapper for creating character objects.

Lang
Wrapper for creating language objects.

List
Wrapper for creating list objects.

ListIter

Logical
Wrapper for creating logical vectors.

StrIter

Symbol
Wrapper for creating symbols.

VecIter

Enums

Robj
Wrapper for an R S-expression pointer (SEXP).

Constants

FALSE

NULL

TRUE

Functions

append⚠

append_with_name⚠

baseenv

end_r

globalenv

make_lang⚠

start_r

Type Definitions

AnyError