Enum rapi::Robj

pub enum Robj {
    Owned(SEXP),
    Borrowed(SEXP),
    Sys(SEXP),
}

Variants

Owned(SEXP)

This object owns the SEXP and must free it.

Borrowed(SEXP)

This object references a SEXP such as an input parameter.

Sys(SEXP)

This object references a SEXP owned by libR.

Methods

impl Robj

pub unsafe fn get(&self) -> SEXP

Get a copy of the underlying SEXP. Note: this is unsafe.

pub unsafe fn get_mut(&mut self) -> SEXP

Get a copy of the underlying SEXP. Note: this is unsafe.

pub fn sexptype(&self) -> u32

Get the XXXSXP type of the object.

pub fn len(&self) -> usize

Get the extended length of the object.

pub fn as_i32_slice(&self) -> Option<&[i32]>

Get a read-only reference to the content of an integer or logical vector.

pub fn as_f64_slice(&self) -> Option<&[f64]>

Get a read-only reference to the content of a double vector.

pub fn as_u8_slice(&self) -> Option<&[u8]>

Get a read-only reference to the content of an integer or logical vector.

pub fn pairlist_iter(&self) -> Option<ListIter>

pub fn list_iter(&self) -> Option<VecIter>

pub fn str_iter(&self) -> Option<StrIter>

pub fn as_str(&self) -> Option<&str>

Get a read-only reference to a char, symbol or string type.

pub fn eval(&self) -> Result<Robj, AnyError>

pub fn eval_blind(&self) -> Robj

pub unsafe fn unprotected(self) -> Robj

pub fn is_owned(&self) -> bool